refactor(api): move (or delete if existing) certification-centers acceptance tests in certification
